+++ package.json
@@ -53,7 +53,7 @@
     "@babel/preset-env": "7.10.4",
     "@babel/preset-react": "7.10.1",
     "@codaco/eslint-plugin-spellcheck": "0.0.14",
-    "@codaco/ui": "~5.4.1",
+    "@codaco/ui": "~5.5.0-alpha.11",
     "@zippytech/sorty": "^2.0.0",
     "ajv": "^6.5.4",
     "animejs": "^2.2.0",
