+++ src/styles/containers/_data-export-screen.scss
@@ -1,0 +1,82 @@
+.data-export-screen {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  position: relative;
+
+  .steps {
+    flex: 1;
+    display: flex;
+  }
+
+  &__close {
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+    width: 2rem;
+    height: 2rem;
+  }
+
+  &__step {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+  }
+
+  &__main {
+    flex: 1;
+    background: linear-gradient(#404088, var(--color-cyber-grape));
+  }
+
+  &__footer {
+    padding: unit(2) unit(6);
+    background-color: var(--color-neon-coral);
+    display: flex;
+    justify-content: space-between;
+
+    .action-buttons {
+      display: flex;
+      align-items: center;
+      justify-content: flex-end;
+
+      .button {
+        &:last-child {
+          margin-left: unit(2);
+        }
+      }
+    }
+  }
+
+  &__session-select {
+    > header {
+      padding: unit(4) unit(6) 0;
+    }
+
+    .content-area {
+      padding: unit(1) unit(6);
+    }
+
+    .selection-status {
+      margin: unit(2) unit(4);
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      flex-direction: row-reverse;
+      height: unit(6);
+
+      .header-toggle {
+        margin-left: 2rem;
+      }
+    }
+
+    .new-filterable-list {
+      margin: unit(4);
+      margin-bottom: 0;
+
+      &__main {
+        max-height: 45vh;
+        min-height: 16rem;
+      }
+    }
+  }
+}
